Fix compilation error.
